input: fix handling of 'forward' in dependencies
The calculation of environment and tools was broken in conjunction with
the usage of forwarded dependencies. The right way to do this is to
strictly keep the environment/tools for the local package and for the
dependencies separated. The local env/tools are always updated and the
others only if 'forward' is True.