input: refactor sandbox and global paths
The sandbox feature was integrated quite inflexible. The project had to
supply a root file system in config.yaml that was handled through some
special logic to download, extract and use it. The sandbox could not be
used in development build mode and caused no rebuild if it changed.

The new schema makes the sandbox a first class citizen. A recipe might
provide itself as a sandbox through 'provideSandbox'. This declares the
$PATH when using it and optionally a couple of mounts from the host into
the sandbox. Other recipes might use and forward the sandbox.

A sandbox is declared the following way:

    provideSandbox:
        paths: ["/bin", "/usr/bin"]
        mount:
            - "/etc/resolv.conf"

while the use of that looks like this:

    depends:
        -
            name: sandbox
            use: [sandbox]

The current sandbox is passed like environment variables or tools
through the dependency tree. Different recipes can thus have different
sandboxes or override the inherited sandbox from upstream recipes.

At the same time the 'globalPaths' option is discarded. The base $PATH
value is defined by the sandbox and on builds without a sandbox the host
$PATH is used. This makes much more sense and is what the user actually
expects.