build: hash build step workspace in development mode
Normally the build step result is recorded just by the date rather than
the content. This makes sense in release mode as the build step ought to
be deterministic. But in development mode the user might change
something in the build step workspace and expect Bob to pick that up.