jenkins: support shared packages
Shared packages are copied to $JENKINS_HOME/bob/.. and directly used
from there for the various jobs. To support slaves the copy-artifact
step is still there to copy in the shared package but it is executed
conditionally. Only if the package is missing it is copied to the
workspace and installed in the shared location.

This finally enables the usage of big toolchains on Jenkins without
having to resort to global installation. Previously the toolchain would
be copied to every workspace. Marking the toolchain as a shared recipe
enables properly tracking it as recipe without wasting too much disk
space (and time) on Jenkins.