input: make digest calculation more robust
Some fields in the digest calculation are variable sized. If such fields
are adjacent (such as the environment) some changes could go undetected.
We're now prefixing these fields with their length to detect any changes.

This was hopefully the last time we had to change the digest calculation
in a backwards incompatible way.