input: add properties and state trackers for plugins
A plugin may define any number of properties and state trackers.

A property describes a key in a recipe that is parsed. The class
handling the property is responsible to validate the data in the recipe
and store the value. The property class handles to inheritance between
recipes and classes too.

A state tracker is a class that is invoked on every step when walking
the dependency tree to instantiate the packages. The state tracker thus
has the responsibility to calculate the final values associated with the
properties for every package. Like properties there can be more than one
state tracker.