input: cache parsed recipes
Up to 90% of the time spent loading the recipes is taken by the Yaml
parser. To speed this up a cache is introduced that holds the parsed
python objects. The stat information of the recipe is used to detect
changes that force a re-parse of the recipe.

Given 1000 moderately complex recipes a 'bob ls' took roughly 3s:

$ time bob ls > /dev/null

real    0m3.300s
user    0m3.268s
sys     0m0.028s

After this commit most of this time is gone:

$ time bob ls > /dev/null

real    0m0.460s
user    0m0.428s
sys     0m0.028s

That's almost back into the ballpark of interactive usage.