jenkins: break cyclic job dependencies if needed
Normally the Jenkins job name is derived from the recipe name. If
multiple packages are coalesced into the same job the dependency graph
might get cyclic. If such cycles are found the package name is used for
affected jobs. If this still leads to cycles the package name with an
incrementing suffix is used.