jenkins: store shared packages by Variant-Id
Previously shared packages were stored under their Build-Id. The problem
is that the Build-Id does not reflect the usage of a sandbox. As such,
the same shared artifact would be used if the sandbox is enabled or
disabled.

Instead use the Variant-Id. This is still sufficient because shared
packages have to be deterministic.