input: refactor 'provideDeps' keyword handling
Previously provided dependencies were realized by literally taking the
entry from the 'depends' section and pass this upwards in the dependency
tree. Though duplicates are removed upstream and some sanity checkes
were performed this could lead to nasty effects.

The new implementation will pass the actual package upwards. Upstream
packages will only get the result of the provided dependency. If the
same package is provided by multiple dependencies to an upstream recipe
the variant-id is checked to see if these are compatible. Issue an
exhaustive error message if the packages are incompatible.