input: provideDeps should be always transitive
A dependency that is passed upwards via provideDeps is passed with all
transitive provided packages. The use clause in the depends section is
there to specify the local usage of the dependency artifacts only. This
way it is possible to define pure virtual packages that have no result,
consume no results and just provide dependencies to their upstream
consumers.