jenkins: support CSRF protection
The Jenkins server might be configured to "Prevent Cross Site Request
Forgery exploits". In this case Bob has to obtain a CSRF protection
token and send it with every POST request. If the server has this
feature disabled the request for a token will 404, which is silently
ignored by Bob.

Fixes #27.