build: fix SCM directory tracking if checkout fails
Previously the directory state of a checkout step was only recorded
_after_ the checkout script returned successfully. This lead to the
problem that the following cycle did not move the SCM directory into the
attic as expected:

  * $ bob dev ...
  * SCM checkout fails but directory was already created (e.g. git
    trying to check out unknown tag)
  * edit recipe and change this very SCM spec
  * try again

The solution is to record an intermediate state for the checkout step
that reflects the anticipated directory state. The script state itself
is not recorded to always force a re-execution of the checkout step in
case anything goes wrong. Once the checkout script succeeded the real
state is recorded.