misc: introduce new query-scm command
This command is used to get the configuration of all SCMs of a package.
The output format can be chosen freely for each SCM type.

This command is also the first of so-called "script commands". These
commands are primarily targeted for scripts and should have a long term
stable output format.