clean: do not prune source directories by default
It is usually not a good idea to remove unused source directories as they
might contain uncommitted changes. Instead let the user explicitly state
that he knows what he is doing.

This change requires an upgrade to the bob state to persist if a
particular directory belongs to a checkout step or not. For upgraded
workspaces the feature will not work because this information is
missing.