input: add {checkout,build,package}VarsWeak
Normally every added environment variable of a step (e.g. buildVars) is
considered significant for the build result. That is, they contribute to
variant management and/or cause a rebuild of a package if the content
changes.

Sometimes there are variables that change behaviour but do not influence
the actual result, though. A typical example is to specify the number of
parallel make jobs. For such use cases variables might be added as
"weak" to a step. This tells Bob that the content of the variable does
not change the (binary) result of the step.