input: make digests of included files long-term stable
The calculation of variant- and build-id must be long-term stable and
must not depend on any implementation details. This requirement was
violated by the $<<>> include syntax. There, the actual method of
including the file was taken into account.

Now there are always two variants of a script: the actual one that gets
executed (as returned by getScript()) and a list of SHA1-hashes
separated by newlines (returned by getDigestScript()). By collapsing the
content into hash sums there is an additional benefit that the build-id
calculation on the Jenkins back end produces considerably smaller shell
code.