Make file modes predictable
For reliable file modes the umask(2) must be the same on all systems.
This has already been enforced inside the sandbox but not outside of
it. As the file mode is also part of the Build-ID for indeterministic
checkouts we have to make it the same everywhere or else binary
artifacts won't work reliably.

On Jenkins the umask is configured from outside. Bob has no means to
change it before the SCM-plugins do their checkouts. Instead we check
for the right mask and bail out if the requirement does not hold.