build: extend --destination option
Originally the --extend option was quite dumb. It just copied the result
of a single package to a destination directory. Before copying it was
erased completely. Additionally the destination option did not work with
multiple packages at all.

The new behaviour allows multiple packages and their result is copied
incrementally to the destination directory. Existing files will be
overwritten but the content will not be erased anymore. Furthermore any
provided dependencies of the built packages (provideDeps) will be
copied too. This is what almost all users are expecting.