input: allow recursive dependency declaration
Instead of naming another recipe directly a dependency entry might
include a list of dependencies. These dependencies will inherit all
attributes that are set in the entry. For example this can be used to
elegantly guard a whole dependency list with an "if" condition:

  depends:
    - if: "condition"
      depends:
        - foo
	- bar
	- baz

This can be nested to any depth. All "if" conditions in the hierarchy
must match. Any other attributes (use, forward, environment) can be
overridden on each level.