audit: introduce audit trail feature
For every artifact an audit trail record is stored tracking the complete
history of all dependencies that contributed to this artifact. This
includes the state of SCMs, the real environment when executing the
build script and the Bob version and recipes that were used for building
the artifacts.

The format of binary artifacts is changed to hold the audit trail next
to the actual content. The result is moved to the 'content' directory
and the 'meta' directory holds the audit trail (audit.json.gz).