test: add test showing directory swap problem
In develop build mode the directories are not 100% stable. If the
recipes changed the directories of the direct dependencies of a build
step might change. This must lead to a clean build because many tools
(e.g. cmake) assume stable input directories on incremental builds.

The test does incremental builds where the first build will store the
directories to the dependencies and subsequent builds rely on the
persisted paths.