build: fix directory changes in develop mode
Directories are reused in develop build mode and thus might change even
though the variant id of a build step is stable. As most tools rely on
stable input directories we have to make a clean build if any of the
dependency directories change.

This problem cannot happen in release build mode because directories
are never reused there.