build: make develop mode directories more stable
Previously the calculation of the develop mode directories was based on
the variant-id of the traversed packages. That has the unwanted side
effect that the order of the traversal may change due to unrelated
changes in the recipe tree.

Now the traversal is done exactly in the order of the recipes.
Directories will now change only if the order of dependencies is
changed or new variants are added and/or removed.