Replaced the old StringParser with a new one.
The new StringParser is actually a suite of parsers and tokenizers.
Which parser and tokenizer is used depends on the actual context.
The behaviour is controlled by rules configured within each tokenizer.
After the tokenizer has finished, the corresponding parser checks, if
the detected tokens match the grammar of the actual context.  If not, a
ParseError with the detected tokens and their textual values is raised.

The hierarchical structure allows solid nesting of different contexts.
If required, tokens and grammar of each context can easily be modified.
The evaluation of the parse result is done in a seperate function.

This new implementation adds some more parameter expansion capabilities:

 * substring removal (%, %%, #, ##)
 * search and replace (//pattern/string, /pattern/string)
 * case modifications (^/^^/,/,,)

Within the patterns character classes ([...]) and wildcards (*, ?)
can be used. Variable substitution is also done within patterns.