archive: support multiple backends simultaneously
Starting with this commit it is possible to specify a list of archive
backends. The usage of each entry can be restricted with an optional
"flags" property ([download, upload] or [download] or [upload]).

If multiple download backends are available they will be tried in order
until a matching artifact is found. All available upload backends are
used for uploading artifacts. Any failing upload will fail the whole
build.