misc: fix query-scm recursive operation
The package tree must be fully traversed in any case. Only the print out
of the SCMs may be skipped to suppress duplicates. Previously also the
recursion was stopped if the checkout step was seen once. This did
prevent the printing of packages lower in the dependency tree under
certain circumstances.