jenkins: add option to keep obsolete jobs
Previously obsolete jobs were always deleted. This had the unpleasant
effect of throwing away all logs and builds of these jobs even if the
recipe was only disabled temporarily. This commit adds a 'keep' option
to retain these jobs and instead disable them. The option is off by
default to keep the original behaviour.

At the same time the job configuration is tracked more fine grained by
Bob: the last set configuration, the last scheduled configuration and
the enable/disable status. This will make the usage of

 $  bob push --no-trigger ...

safe as Bob knows that it still has to schedule these jobs when being
run again.