scripts: handle broken pipes
If the output of bob if piped trough a pager (e.g. 'head' or 'less') it
is not uncommon that stdout closes before all output is written. In this
case a BrokenPipeError is thrown. Catching this exception is not enough,
though. Unless stderr is closed too the following message is still
shown:

 Exception ignored in: <_io.TextIOWrapper name='<stdout>' mode='w' encoding='UTF-8'>
 BrokenPipeError: [Errno 32] Broken pipe