input: reuse common package trees
For the most part this commit restructures the internal data structures.
It splits off the location independent data structures (as in "package
path") into separate Core{Step,Package,...} classes. These core classes
represent the tree in a "relative" fashion. The user visible, location
dependent classes are created on the fly while the tree is traversed.

The second part of the change is to use the deterministic calculation to
detect already calculated packages. For that the input to the package
calculation with the used tools and environment in the sub-tree is
compared with already created packages. Reusing these common sub-trees
dramatically brings down memory consumption and shrinks the package
cache considerably.