ls: show correct indirect dependencies
By default only dependencies that are named in the recipe are shown. By
adding '-a' the ls command also shows the indirect dependencies, that is
all provided dependencies that were passed upwards via 'provideDeps'.

Previously the '-a' switch showed all dependencies (including tools and
sandbox) which is wrong (these packages typically come from somewhere
else in the tree) and is not useful.