build: support multiple directories per variant
The build logic was using the variant-id to track what has been built
already. This breaks if different recipes with the same variant-id are
mapped to different directories. Instead we now track the built
directories because a workspace directory must always map to a single
variant-id.

Fixes a regression introduced by aa9c9fea that added the possibility
that different directories are used for the same variant-id.