build: harden against interrupted/failed steps
If the execution of a step fails it will leave the affected work space
in an undefined state. This has two consequences:

1. The step should always run again if we reference it in future
   invocations.
2. Dependent steps must know that the result of the current step has
   changed.

The strategy to address these problems is to remove the stored input
hashes of the step and forge the result hash before executing the step.
This will make sure the step will always run again. Any dependent steps
will see the forged result hash (a time stamp) and will run again too.