build/jenkins: always hash sources for build-id
When working with "bob dev" the developer expects to be able to change
any of the involved sources, call bob again and it will rebuild with the
changes. If binary artifacts are used and the sources are deterministic
this was not true anymore. The problem was that Bob treated
deterministic sources different. For these the actual sources were
irrelevant and the binary artifact was downloaded just by looking at the
configured tag/commit.

Starting with this commit Bob doesn't treat tags/commit special anymore.
Bob will always checkout all involved sources first, create a hash sum
and use that as the access key for the binary artifact. This has already
been the behavior for checkouts from branches.

As this changes the calculation of the build-id any artifacts that were
built from tags will be incompatible with previous versions of Bob.

Fixes #106.