build: fix flaky handling of "already executed" memory
Due to historical reasons the "was already run" map was implemented as
bijective mapping. This is neither required nor correct anymore. Since
86bdac6f (build: support multiple directories per variant) it is
possible that multiple directories hold the same variant-id.

The old implementation thus could lead to situations where a "this step
was already run" entry got evicted falsely. This lead to the incorrect
re-execution of steps, increasing the build time.