jenkins: always consider sandbox
The Build-Id is calculated incrementally by each job. Because the
Build-Id calculation forcefully includes the sanbox we have to make sure
that a proper job is created and included. This must be done even if the
sandbox image is not used or else the bob-hash-engine will stop on some
missing input.

Fixes #139.