pathspec: add query language
The query language is backwards compatible to the old path syntax. It
massively expands on the possibilities to match multiple packages,
though.

Parsing of the complex queries is done by the pyparsing module.
Internally the evaluation is done with sets of packages. Each step of
the path evaluation takes the currently selected package set, applies
the axis and predicates and yields the new package set. Predicate
expressions are done as set operations on all packages. Inside of
predicates paths are evaluated backwards to yield the possible set of
packages that match the path.

See "Efficient algorithms for processing XPath queries" [1] for the core
algorithms that are applied here.

[1] Gottlob, G., Koch, C., and Pichler, R. 2002. Efficient algorithms
    for processing XPath queries. In Proceedings of the 28th
    International Conference on Very Large Data Bases (VLDB'02).
    HongKong, China, 95-106