input: refactor path parsing
Split the traversal of the package graph from the parsing module. The
parser has grown way too big. Moving the path traversal code to a
separate module separates these independent steps into distinct modules.

Also sanitize the alias handling in preparation of advanced queries. The
alias is applied to the first part of a path. Do it as a string
substitution because the alias might yield more than one package in the
future.