pathspec: add 'direct-' axis
Normally all dependencies between packages are considered when
evaluating the location step axis. This ensures that any possible path
is accepted by default, especially in the abbreviated format. The
'direct-child', 'direct-ancestor' and 'direct-descendant-or-self' axis
can be used to restricted the step to direct paths only. Use these axis
if the query should not consider provided dependencies.