archive: add support for live-build-ids
Live-build-id to build-id translations are cached on the binary archive
as small files that use the live-build-id as access key and hold the
real build-id as content. The archive backends just need to store and
retrieve these small files. They are treated more or less like the real
artifacts and share the same folder structure.