build: add live-build-id support
Extend the build logic to query the live-build-id if possible and use
the archive to translate this into a real build-id. If this succeeds a
proper binary artifact is searched. The last queried live-build-id is
cached for subsequent "build-only" builds. For regular builds the
live-build-id is always queried. The live-build-id to build-id
translation is cached too.

For the upload of the live-build-id to build-id translation only virgin
checkouts are considered. The local user may tamper with the checked out
sources which we cannot reliably detect. Choose the safe side and only
upload where we can be sure that all is unmodified.

The build logic has to handle an additionl corner case that didn't exit
previously. The live-build-id and the checked out sources do not need to
be in sync necessarily. If there is a check-in on the SCM in between
these two operations the predicted build-id and the real one may not
match. Handle the case by invalidating the affected and any possibly
derived build-ids and by restarting the build to re-examine all build-
and package-steps again.