build: pre- and post-scripts
Add two optional commands to default.yaml et.al. that can be run before
(preBuildHook) and after (postBuildHook) the build. The preBuildHook
receives the packages that should be built. It will fail the build if it
returns with a non-zero status. The postBuildHook will get the status
(success/fail) and the paths to the results as arguments.