Add optimistic mode to drop the fsync.
bob will quite often flush to disk using `fsync`. This overhead becomes
significant when using small projects and a rotating hard disk.  The
protection it provides is comparatively minor, because if the system
actually crashes, object files created by the compiler will be in an
indeterminate state anyway; the compiler does not do `fsync`.  Thus, this
adds an option `--optimistic` to disable the `fsync`.