jenkins: rewrite job name calculator
The old job names were determined optimistically by first putting all
packages derived from a recipe into the same job. Then the job graph was
checked for cycles and we began to split jobs as we found them
iteratively.

This proves to be suboptimal for recursive recipe structures. Instead we
now first put every package in a separate job and then try to coalesce
them based on their recipe name. This is done as long as no job cycles
are produced by the merges.

If the recipe name cannot be used we try next with the longest common
prefix of all packages in a job. This usually works and leads to natural
job names based on the multiPackage entries in a recipe. As a last
resort we fall back on a counting suffix as the previous algorithm to
disambiguate jobs.