bob: improve error diagnostics
When executing the final scripts they are a concatenation of the
inherited classes and the recipe script itself. If an error happens the
line number of the failing command is pretty much useless to the user
because there is no connection to the source.

Starting with this commit the error handler will print a backtrace. For
every entry of the backtrace the function name and the line number in
the originating class/recipe script is printed. This looks like this:

    Step failed with return status 127; Command: fn4 does-not-exist
    Call stack (most recent call first)
	#0: Class  bar, line 4, in fn3
	#1: Recipe root$1, line 3, in fn2
	#2: Recipe root-foo$1, line 3, in fn1
	#3: Recipe root-foo-baz, line 1, in main