jenkins: prevent running in stale workspaces
Jenkins does not delete workspaces automatically. He also has no clue if
the job that is running in an old workspace is compatible or not. Hence
we have to make sure ourself that we do not collide with old data in
case of incremental builds.

To pro-actively prevent such clashes every Jenkins server in Bob now
get's a 32-bit random UUID which is appended to every path. This should
minimize the chance to collide with another or old instance.

Because of the low entropy the UUID does not provide a guarantee that no
collision occurs. It also does not help on existing instances where no
UUID was assigned. Therefore, as a 2nd line of defence, every workspace
stores a canary with the variant-id. If the canary exists it has to
match. Otherwise the build is aborted because we don't build the same
thing in this existing workspace. The user has to clean up the workspace
manually then.

Fixes #156.