build: fix postBuildHook status
We incorrectly set the success status already to "True" after the first
successfully built package. But if we build more than one package and
fail later then the complete build has failed but the postBuildHook was
called with $1 == "success".

Now we return success only after the last package was built
successfully.