input: filter correctly if dependency is named multiple times
It is possible (but not useful) to name the same dependency multiple
times in a recipe. If only the tools are used it was not detected when
multiple variants of the same package were specified. We now do the
duplicate check independent of what is actually used.

If a package is named multiple times in the depends section of a recipe
and they are the same variant then we will issue a warning. It is (and
was already in the past) defined that only the first in the list is
taken. But it is most likely an error of the user if there is more than
one reference to the same package. It is also possible that, even if the
packages themself are of the same variant, they might provide different
dependencies or variables upwards. This is still not detected and will
be prohibited implicitly by turning the warning into an error in the
future.

Fixes #174.