input: improve duplicate package matching runtime
Instead of searching a linear list of package candidates we try to
create a tree of computed packages. The root of the tree is always a
PackageList object. It does not directly hold the data but references a
PackageMatchList list at the beginning that holds the actual package
match entries. If the number of packages grows we try to form a tree
where intermediate nodes are comprised of PackageDiscriminator instances
that do selective search based on a certain attribute (env key, tools
key, sanbox). The leaves of the tree are always PackageMatchList
objects.

Also do not copy environment and tools for comparison every time.
Instead use a read-only view onto the data with Env.inspect().