input: compare PluginState by value
While traversing the recipe dependencies and preparing the packages we do
create a new PluginState objects at every step. To see if they have
changed we have to compare their *content* instead of their *identity*
which will always be different.

We now provide a reasonable default implementation that compares every
member of the object. A plugin may overwrite them with a more sensible
approach if necessary.