url: tidyUrlSCM: track checkout directory instead of file
The URL SCM was not tracking the checkout directory but the individual
files that are downloaded by the SCM. This has the advantage that it is
possible to download more than one file into the same directory. There
are a couple of major disadvantages, though:

1. When extracting multiple archives in the same directory it might be
   possible that some files are overwritten.
2. Any extracted files are not tracked by Bob and will be left untouched
   in develop mode when the recipe is updated. This leads to stale files
   in the src-directory and will typically prevent that matching binary
   artifacts are found.
3. Trying to reliably apply patches across SCM updates is tricky because
   files are only overwritten and never garbage collected.

Starting with this commit Bob will manage the whole checkout directory. This
unifies the behaviour with the other SCMs and solves the above
disadvantages. This change might break existing projects because with
the new behaviour it is not possible to put multiple URL SCMs into the
same directory. The new behaviour is thus controlled by the tidyUrlSCM
introduced by this commit.