build: allow build workspace to symlink checkout result
This commit adds support for having symlinks to files of the checkout
workspace in the build workspace (e.g. linking big archives). It has
always been guaranteed to have the checkout and build workspace of the
package available in the package step. Thus the package step can access
this link instead of having to copy the archive in the build step.

Instead of hashing the build workspace differently we add the checkout
step to the checklist of the package step. It also better reflects the
visibility of the checkout workspace during the package step.