archive: add azure backend
The azure backend uses the Microsoft cloud block blob storage. It will use the
same layout as all other backends so that the container can be made
public readable and be fetched by the http backend without giving out
the credentials.

Given a typical configuration like this:

  archive:
      backend: azure
      account: <account>
      container: <container name>
      key: <access key>

a publicly readable container can also be used by:

  archive:
      backend: http
      url: https://<account>.blob.core.windows.net/<container name>
      flags: [download]