input: improve error message on invalid policy
During development of the next version we might already introduce new
policies. These policies are not known to older versions but unless the
first release candidate is tagged it is not yet possible to force a
recent version by bobMinimumVersion. So an unknown policy is probably
just an indication that the current Bob version is too old.