bob: replace dbm with sqlite3
The dbm module proves to be very flaky. If interrupted at the wrong time
it even has the ability to subsequently crash the Python VM while
loading the database. It also depends on the available implementations
and, if the distro changes it across updates, might break existing
workspaces. It doesn't support concurrent access either.

Let's hope that sqlite3 is better in all the above cases. At least it's
not slower and has a built-in transaction and concurrency model.

Fixes #118 and #183.