build: check all dependencies for input changes
Traditionally only the arguments were checked for changes when
determining if a step should be run again or not. This is not really
desirable because when working on tools it will not lead to re-builds of
packages that depend on this tool. By taking all dependencies into
account we can properly support these cases too.

Fixes #173.