input: remove scm merging
Merging SCMs at package generation time turned out to be a classic
example of premature optimization. This is only a minor improvement for
Jenkins by collapsing multiple SVN checkouts into one instance but makes
the SCM API needlessly complex.

If this feature is ever needed again it can be implemented in the
back-end as an optimization pass.