input: define relocatable property
Historically any recipe that did define at least one tool was deemed to
be non-relocatable. This implied that it cannot be stored as binary
artifact unless we're in a sandbox.

This is overly pessimistic. By introducing a 'relocatable' property the
user is able to express if a package is really relocatable or not. For
backwards compatibility the default value is False if the recipe
defines a tool, otherwise it is True.