build/jenkins: use isRelocatable instead of doesProvideTools heuristic
The only reason why artifacts should not be up- or downloaded is if they
are not relocatable. The heuristic to treat all tools as non-relocatable
is overly pessimistic and by the introduction of the 'relocatable'
recipe property the user is able to express this directly.

Any other side effects on tools that are executed on the host can be
expressed properly in the recipes. The recommended course of action is
that the tools are linked statically and the host architecture is passes
as environment variable. If host libraries are used then the recipe
should be marked as not relocatable.