input: add plugin settings support
Sometimes plugin behaviour needs to be configurable by the user. On the
other hand Bob expects plugins to be deterministic. To have a common
interface for such settings it is possible for a plugin to define
additional keywords in default.yaml. This provides Bob with the
information to validate the settings and detect changes in a reliable
manner.