build: improve clean-build heuristic
The regular heuristic that detects the need of clean builds is based on
the variant-id. Whenever the variant-id changes compared to the
previous build in the same directory we make a clean build. By and
large this heuristic has proven to be correct.

Unfortunately the user might build single packages, skip dependencies
or do build-only builds. In this case the calculated variant-id of the
recipes does not really match the variant-id of what is built. This
leads to the following bugs:

 * Just the checkout is changed but the user builds with --build-only.
   In this case we did a clean build which is unnecessary because no
   checkout was done.
 * If the user then runs again without --build-only we do *not* do a
   clean rebuild. This was done in the previous run but for the wrong
   sources.

We now calculate the variant-id incrementally based on the last runs in
the workspaces. After every step the incremental variant-id is stored
for the workspace. The next step uses these stored variant-ids as
input. This is much closer to reality.

Fixes #198.