input: implement sandboxInvariant policy
Traditionally the impact of a sandbox to the build has not been handled
consistently. On one hand the actual usage of a sandbox was not relevant
for binary artifacts. As such, an artifact that was built inside a
sandbox was also used when building without the sandbox (and vice
versa). On the other hand Bob did rebuild everything from scratch when
switching between sandbox/non-sandbox builds. This inconsistent behavior
is rectified by the sandboxInvariant policy that consistently declares
builds as invariant of the sandbox.