input: optimize members of Step classes
Everything that is constant for a recipe or for a package should be
stored there and not replicated for every Step object. This reduces the
amount of data that must be persisted across invocations considerably.

As runtime optimization we do not create the whole script for 'isValid'
in case of checkout steps to just see that it is not None.