build: fix path name calculation
There was a slight regression that got introduced in d31f9541 which
triggered an assertion that the workspace path was not in the database
even though it should have been there. The problem was that the
packages must be traversed on the package level to visit all packages
that are reachable by the query language.