build: fix crash when updating recipes
When updating recipes the still matching directories were taken over.
Starting with a00308d this could lead to the following error:

  Error: Cannot save directory mapping: UNIQUE constraint failed: dirs.key

The problem was that the directories that were taken over from the old
database were not marked as "visited". If the same key also mapped to a
different directory (e.g. a common build step of multiPackages) then the
algorithm would try to add the same key multiple times.

Now consistently mark each key as visited and process each only once.

Fixes #201.