build/jenkins: take exec path into account
Most build systems fail when the workspace is moved. Because of the
buildInvariant policy it is now possible to have incremental builds in
and outside the sandbox. From the perspective of the package build step
this looks like the workspace was moved. Do a clean build instead if the
exec path changes compared to the last build in the same workspace.

This is irrelevant for package steps because they are always built
clean. Checkout steps are also not regarded as affected because they
cannot be built clean anyway.