input: add 'environment' property to 'provideSandbox'
A sandbox brings its own host environment that overrides the host
environment when used. This can be quite different and be even a
different architecture (think of an i386 sandbox image that is used on
an x86_64 host).

Bob regards the sandbox as a build invariant, that is it must have no
impact on the results of the recipes that are built inside the sandbox.
On the other hand there can be situations where this still the case. To
properly handle this a sandbox recipe can now define environment
variables that are picked up when this sandbox is actually used. By
consuming this variable in other recipes Bob can determine the impact of
the sandbox, if any.