input: allow user to amend sandbox mounts/paths
The default sandbox mount and search paths are defined in the sandbox
recipe(s). If the user wanted to mount additional directories inside the
sandbox (e.g. a local source code mirror in conjunction with a
scmOverride) he previously had to edit the recipes.

Now it is possible to add such paths to default.yaml.