jenkins: warn about non-relocatable packages
The Jenkins build backend needs to copy packages between workspaces.
This is especially important if multiple build slaves are used but also
to keep jobs independent of each other. If a non-relocatable package is
built outside of a sandbox then this might fail, even if the recipe
marked it as "shared". Warn the user if such a configuration is
detected.