input: fix missing / duplicated indirectPackages
With 38bc5df1 the behavior of the indirectPackages filter has been
changed. When a dependency's result is not used (use: []) but the same
dependency is picked up as indirect package then the result was not
picked up as expected. That was working before 38bc5df1 and is restored
by this commit.

OTOH it could happen that the same result was picked up more than once.
Even the original algorithm could show this behavior. Starting with this
commit results are always picked up exactly once. Note that this might
lead to incompatible binary artifacts because the variant-id might
change for some packages.

This also fixes a possible incorrect provision of a package via
'provideDeps' of a disabled dependency. If the dependency was disabled
due to a failing 'if' condition but was also picked up as an indirect
dependency then it would falsely be provided. Before 38bc5df1 it would
not be provided. This commit restores the old, correct behavior.